src/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
openpyxl==3.1.2
fpdf==1.7.2
langgraph==0.0.20
langchain-openai==0.0.2
langchain-core==0.1.5
python-dotenv==1.0.0
streamlit==1.28.0
pandas==2.1.1
requests==2.31.0
